German translation: nondiscriminatory language
The original German translation explicitly addresses male players, which
in conservative reading of German language is meant to address everyone.
The changes in this commit replace this by a gender-neutral language.
There is no universal consensus in Germany concerning gender-neutral
language and some conservative readers might prefer the "traditional"
wording. But discrimination-free language is no longer just a political
topic but also part of everyday live in Germany.

Changes also address a grammar mistake in the German translation
(line 99: "unschuldigen" instead of "unschuldiger").